@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 
-Widget ncButton(String buttonText, {@required VoidCallback onPressed}) {  
+Widget ncButton(String buttonText, {
+    @required VoidCallback onPressed,
+  }) {  
   return FlatButton(
     child: Text(buttonText),
     onPressed: onPressed,
