@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-void showDialogSingleButton(BuildContext context, String title, String message, String buttonLabel) {
+void showDialogSingleButton(BuildContext context, String title, String message, {String buttonLabel = 'OK'}) {
   // flutter defined function
   showDialog(
     context: context,
